docs: consolidate example processing
