+++ src/gas/examples/e2eTest01.js
@@ -1,8 +1,11 @@
+import EnvContext from "../../lib/Context"
+import * as GmailProcessorLib from "../../lib/index"
+
 /**
  * This is a special configuration used for end-to-end testing using the emails generated by `e2eInit.js`.
  * @type {GmailProcessorLib.Config}
  */
-const e2eTest01Config = {
+export const e2eTest01Config = {
   description: "End-to-end (E2E) test configuration",
   settings: {
     logSheetLocation:
@@ -134,7 +137,7 @@ const e2eTest01Config = {
   ],
 }
 
-function e2eTest01Run(
+export function e2eTest01Run(
   /** @type {EnvContext | undefined} */
   ctx,
 ) {
