+++ src/gas/examples/exampleMin.js
@@ -1,10 +1,13 @@
+import EnvContext from "../../lib/Context"
+import * as GmailProcessorLib from "../../lib/index"
+
 /**
  * This is a minimal configuration example, just with the required information.
  * It just processes files without doing anything, but they will still be marked as processed!
  * It can be used if you want to start your config from scratch.
  * @type {GmailProcessorLib.Config}
  */
-const exampleMinConfig = {
+export const exampleMinConfig = {
   settings: {
     // Chose a method to mark processed entities:
     markProcessedMethod:
@@ -22,7 +25,7 @@ const exampleMinConfig = {
  * @param {EnvContext | undefined} ctx Environment context
  * @returns {GmailProcessorLib.ProcessingResult} Processing result
  */
-function exampleMinRun(ctx) {
+export function exampleMinRun(ctx) {
   return GmailProcessorLib.run(
     exampleMinConfig,
     GmailProcessorLib.RunMode.DRY_RUN,
