--- src/test/examples/example02.js
@@ -1,74 +1,0 @@
-import EnvContext from "../../lib/Context"
-import * as GmailProcessorLib from "../../lib/index"
-
-/**
- * This is a more advanced configuration example.
- * @type {GmailProcessorLib.Config}
- */
-export const example02Config = {
-  description: "An example V2 configuration",
-  settings: {
-    markProcessedMethod:
-      GmailProcessorLib.MarkProcessedMethod.MARK_MESSAGE_READ,
-    maxBatchSize: 10,
-    maxRuntime: 280,
-    sleepTimeThreads: 100,
-    sleepTimeMessages: 0,
-    sleepTimeAttachments: 0,
-    timezone: "UTC",
-  },
-  global: {
-    thread: {
-      match: {
-        query: "has:attachment -in:trash -in:drafts -in:spam newer_than:1d",
-        maxMessageCount: -1,
-        minMessageCount: 1,
-      },
-      actions: [],
-    },
-  },
-  messages: [
-    {
-      description: "Message shorthand config",
-      match: {
-        subject: "My Subject",
-      },
-    },
-  ],
-  attachments: [
-    {
-      description: "Attachment shorthand config",
-      match: {
-        name: "my-file-.*",
-      },
-    },
-  ],
-  threads: [
-    {
-      description:
-        "Store all attachments sent to my.name+scans@gmail.com to the folder 'Scans'",
-      match: {
-        query: "to:my.name+scans@gmail.com",
-      },
-      actions: [
-        {
-          name: "thread.storePDF",
-          args: {
-            folder: "Scans-${message.date:format:yyyy-MM-dd}",
-          },
-        },
-      ],
-    },
-  ],
-}
-
-export function example02Run(
-  /** @type {EnvContext | undefined} */
-  ctx,
-) {
-  return GmailProcessorLib.run(
-    example02Config,
-    GmailProcessorLib.RunMode.DRY_RUN,
-    ctx,
-  )
-}
