--- src/test/examples/migrationExample01.js
@@ -1,52 +1,0 @@
-import * as GmailProcessorLib from "../../lib/index"
-
-/**
- * This is a Gmail2GDrive v1.x configuration example to demostrate the conversion to the Gmail Processor v2.x config format.
- * @type {GmailProcessorLib.V1Config}
- */
-export const migrationExample01Config = {
-  globalFilter: "has:attachment -in:trash -in:drafts -in:spam",
-  processedLabel: "gmail2gdrive/client-test",
-  sleepTime: 100,
-  maxRuntime: 280,
-  newerThan: "1d",
-  timezone: "GMT",
-  rules: [
-    {
-      filter: "to:my.name+scans@gmail.com",
-      folder: "'Scans'-yyyy-MM-dd",
-    },
-    {
-      filter: "from:example1@example.com",
-      folder: "'Examples/example1'",
-    },
-    {
-      filter: "from:example2@example.com",
-      folder: "'Examples/example2'",
-      filenameFromRegexp: ".*.pdf$",
-    },
-    {
-      filter: "(from:example3a@example.com OR from:example3b@example.com)",
-      folder: "'Examples/example3ab'",
-      filenameTo: "'file-'yyyy-MM-dd-'%s.txt'",
-      archive: true,
-    },
-    {
-      filter: "label:PDF",
-      saveThreadPDF: true,
-      folder: "'PDF Emails'",
-    },
-    {
-      filter: "from:example4@example.com",
-      folder: "'Examples/example4'",
-      filenameFrom: "file.txt",
-      filenameTo: "'file-'yyyy-MM-dd-'%s.txt'",
-    },
-  ],
-}
-
-export function migrationExample01ConvertConfig() {
-  const config = GmailProcessorLib.convertV1Config(migrationExample01Config)
-  console.log(JSON.stringify(config, null, 2))
-  return config
-}
